feat: add --no-hmr flag
