+++ src/client/hmr.ts
@@ -1,0 +1,342 @@
+// This file runs in the browser.
+
+// injected by serverPluginHmr when served
+declare const __PORT__: number
+
+import { HMRRuntime } from 'vue'
+import { HMRPayload, UpdatePayload, MultiUpdatePayload } from '../hmrPayload'
+
+console.log('[vite] connecting...')
+
+declare var __VUE_HMR_RUNTIME__: HMRRuntime
+
+const socketProtocol = location.protocol === 'https:' ? 'wss' : 'ws'
+const socketUrl = `${socketProtocol}://${location.hostname}:${__PORT__}`
+const socket = new WebSocket(socketUrl, 'vite-hmr')
+
+function warnFailedFetch(err: Error, path: string | string[]) {
+  if (!err.message.match('fetch')) {
+    console.error(err)
+  }
+  console.error(
+    `[hmr] Failed to reload ${path}. ` +
+      `This could be due to syntax errors or importing non-existent ` +
+      `modules. (see errors above)`
+  )
+}
+
+// Listen for messages
+socket.addEventListener('message', async ({ data }) => {
+  const payload = JSON.parse(data) as HMRPayload | MultiUpdatePayload
+  if (payload.type === 'multi') {
+    payload.updates.forEach(handleMessage)
+  } else {
+    handleMessage(payload)
+  }
+})
+
+async function handleMessage(payload: HMRPayload) {
+  const { path, changeSrcPath, timestamp } = payload as UpdatePayload
+  switch (payload.type) {
+    case 'connected':
+      console.log(`[vite] connected.`)
+      break
+    case 'vue-reload':
+      queueUpdate(
+        import(`${path}?t=${timestamp}`)
+          .catch((err) => warnFailedFetch(err, path))
+          .then((m) => () => {
+            __VUE_HMR_RUNTIME__.reload(path, m.default)
+            console.log(`[vite] ${path} reloaded.`)
+          })
+      )
+      break
+    case 'vue-rerender':
+      const templatePath = `${path}?type=template`
+      import(`${templatePath}&t=${timestamp}`).then((m) => {
+        __VUE_HMR_RUNTIME__.rerender(path, m.render)
+        console.log(`[vite] ${path} template updated.`)
+      })
+      break
+    case 'style-update':
+      // check if this is referenced in html via <link>
+      const el = document.querySelector(`link[href*='${path}']`)
+      if (el) {
+        el.setAttribute(
+          'href',
+          `${path}${path.includes('?') ? '&' : '?'}t=${timestamp}`
+        )
+        break
+      }
+      // imported CSS
+      const importQuery = path.includes('?') ? '&import' : '?import'
+      await import(`${path}${importQuery}&t=${timestamp}`)
+      console.log(`[vite] ${path} updated.`)
+      break
+    case 'style-remove':
+      removeStyle(payload.id)
+      break
+    case 'js-update':
+      queueUpdate(updateModule(path, changeSrcPath, timestamp))
+      break
+    case 'custom':
+      const cbs = customUpdateMap.get(payload.id)
+      if (cbs) {
+        cbs.forEach((cb) => cb(payload.customData))
+      }
+      break
+    case 'full-reload':
+      if (path.endsWith('.html')) {
+        // if html file is edited, only reload the page if the browser is
+        // currently on that page.
+        const pagePath = location.pathname
+        if (
+          pagePath === path ||
+          (pagePath.endsWith('/') && pagePath + 'index.html' === path)
+        ) {
+          location.reload()
+        }
+        return
+      } else {
+        location.reload()
+      }
+  }
+}
+
+let pending = false
+let queued: Promise<(() => void) | undefined>[] = []
+
+/**
+ * buffer multiple hot updates triggered by the same src change
+ * so that they are invoked in the same order they were sent.
+ * (otherwise the order may be inconsistent because of the http request round trip)
+ */
+async function queueUpdate(p: Promise<(() => void) | undefined>) {
+  queued.push(p)
+  if (!pending) {
+    pending = true
+    await Promise.resolve()
+    pending = false
+    const loading = [...queued]
+    queued = []
+    ;(await Promise.all(loading)).forEach((fn) => fn && fn())
+  }
+}
+
+// ping server
+socket.addEventListener('close', () => {
+  console.log(`[vite] server connection lost. polling for restart...`)
+  setInterval(() => {
+    fetch('/')
+      .then(() => {
+        location.reload()
+      })
+      .catch((e) => {
+        /* ignore */
+      })
+  }, 1000)
+})
+
+// https://wicg.github.io/construct-stylesheets
+const supportsConstructedSheet = (() => {
+  try {
+    new CSSStyleSheet()
+    return true
+  } catch (e) {}
+  return false
+})()
+
+const sheetsMap = new Map()
+
+export function updateStyle(id: string, content: string) {
+  let style = sheetsMap.get(id)
+  if (supportsConstructedSheet && !content.includes('@import')) {
+    if (style && !(style instanceof CSSStyleSheet)) {
+      removeStyle(id)
+      style = undefined
+    }
+
+    if (!style) {
+      style = new CSSStyleSheet()
+      style.replaceSync(content)
+      // @ts-ignore
+      document.adoptedStyleSheets = [...document.adoptedStyleSheets, style]
+    } else {
+      style.replaceSync(content)
+    }
+  } else {
+    if (style && !(style instanceof HTMLStyleElement)) {
+      removeStyle(id)
+      style = undefined
+    }
+
+    if (!style) {
+      style = document.createElement('style')
+      style.setAttribute('type', 'text/css')
+      style.innerHTML = content
+      document.head.appendChild(style)
+    } else {
+      style.innerHTML = content
+    }
+  }
+  sheetsMap.set(id, style)
+}
+
+function removeStyle(id: string) {
+  let style = sheetsMap.get(id)
+  if (style) {
+    if (style instanceof CSSStyleSheet) {
+      // @ts-ignore
+      const index = document.adoptedStyleSheets.indexOf(style)
+      // @ts-ignore
+      document.adoptedStyleSheets = document.adoptedStyleSheets.filter(
+        (s: CSSStyleSheet) => s !== style
+      )
+    } else {
+      document.head.removeChild(style)
+    }
+    sheetsMap.delete(id)
+  }
+}
+
+async function updateModule(
+  id: string,
+  changedPath: string,
+  timestamp: number
+) {
+  const mod = hotModulesMap.get(id)
+  if (!mod) {
+    console.error(
+      `[vite] got js update notification for "${id}" but no client callback ` +
+        `was registered. Something is wrong.`
+    )
+    console.error(hotModulesMap)
+    return
+  }
+
+  const moduleMap = new Map()
+  const isSelfUpdate = id === changedPath
+
+  // make sure we only import each dep once
+  const modulesToUpdate = new Set<string>()
+  if (isSelfUpdate) {
+    // self update - only update self
+    modulesToUpdate.add(id)
+  } else {
+    // dep update
+    for (const { deps } of mod.callbacks) {
+      if (Array.isArray(deps)) {
+        deps.forEach((dep) => modulesToUpdate.add(dep))
+      } else {
+        modulesToUpdate.add(deps)
+      }
+    }
+  }
+
+  // determine the qualified callbacks before we re-import the modules
+  const callbacks = mod.callbacks.filter(({ deps }) => {
+    return Array.isArray(deps)
+      ? deps.some((dep) => modulesToUpdate.has(dep))
+      : modulesToUpdate.has(deps)
+  })
+
+  await Promise.all(
+    Array.from(modulesToUpdate).map(async (dep) => {
+      const disposer = disposeMap.get(dep)
+      if (disposer) await disposer(dataMap.get(dep))
+      try {
+        const newMod = await import(
+          dep + (dep.includes('?') ? '&' : '?') + `t=${timestamp}`
+        )
+        moduleMap.set(dep, newMod)
+      } catch (e) {
+        warnFailedFetch(e, dep)
+      }
+    })
+  )
+
+  return () => {
+    for (const { deps, fn } of callbacks) {
+      if (Array.isArray(deps)) {
+        fn(deps.map((dep) => moduleMap.get(dep)))
+      } else {
+        fn(moduleMap.get(deps))
+      }
+    }
+
+    console.log(`[vite]: js module hot updated: `, id)
+  }
+}
+
+interface HotModule {
+  id: string
+  callbacks: HotCallback[]
+}
+
+interface HotCallback {
+  deps: string | string[]
+  fn: (modules: object | object[]) => void
+}
+
+const hotModulesMap = new Map<string, HotModule>()
+const disposeMap = new Map<string, (data: any) => void | Promise<void>>()
+const dataMap = new Map<string, any>()
+const customUpdateMap = new Map<string, ((customData: any) => void)[]>()
+
+export const createHotContext = (id: string) => {
+  if (!dataMap.has(id)) {
+    dataMap.set(id, {})
+  }
+
+  // when a file is hot updated, a new context is created
+  // clear its stale callbacks
+  const mod = hotModulesMap.get(id)
+  if (mod) {
+    mod.callbacks = []
+  }
+
+  const hot = {
+    get data() {
+      return dataMap.get(id)
+    },
+
+    accept(callback: HotCallback['fn'] = () => {}) {
+      hot.acceptDeps(id, callback)
+    },
+
+    acceptDeps(
+      deps: HotCallback['deps'],
+      callback: HotCallback['fn'] = () => {}
+    ) {
+      const mod: HotModule = hotModulesMap.get(id) || {
+        id,
+        callbacks: []
+      }
+      mod.callbacks.push({
+        deps: deps as HotCallback['deps'],
+        fn: callback
+      })
+      hotModulesMap.set(id, mod)
+    },
+
+    dispose(cb: (data: any) => void) {
+      disposeMap.set(id, cb)
+    },
+
+    // noop, used for static analysis only
+    decline() {},
+
+    invalidate() {
+      location.reload()
+    },
+
+    // custom events
+    on(event: string, cb: () => void) {
+      const existing = customUpdateMap.get(event) || []
+      existing.push(cb)
+      customUpdateMap.set(event, existing)
+    }
+  }
+
+  return hot
+}
