+++ src/node/server/serverPluginCss.ts
@@ -19,7 +19,13 @@ import { clientPublicPath } from './serverPluginClient'
 
 export const debugCSS = require('debug')('vite:css')
 
-export const cssPlugin: ServerPlugin = ({ root, app, watcher, resolver }) => {
+export const cssPlugin: ServerPlugin = ({
+  root,
+  app,
+  watcher,
+  resolver,
+  config
+}) => {
   app.use(async (ctx, next) => {
     await next()
     // handle .css imports
@@ -34,7 +40,9 @@ export const cssPlugin: ServerPlugin = ({ root, app, watcher, resolver }) => {
         ctx.type = 'js'
         // we rewrite css with `?import` to a js module that inserts a style
         // tag linking to the actual raw url
-        ctx.body = codegenCss(id, css, modules)
+        ctx.body = config.hmr
+          ? codegenCss(id, css, modules)
+          : `export default ${JSON.stringify(modules || css)}`
       }
     }
   })
